Finished refactor and fixed fill algorithm along curves
