+++ Hephaestus/src/Torchlight/Core/UICreationToolApplication.cpp
@@ -1,13 +1,13 @@
 
 #include <Torchlight/Core/UICreationToolApplication.hpp>
 
-#include <Torchlight/Util/FontProcessor.hpp>
-
 #include <Platform/Vulkan/VkUtil.hpp> //Hack to get hard coded values working. This will be fixed when vertices are moved to their own file.
 
 #include <HellfireControl/Render/Renderer.hpp>
 
 #include <HellfireControl/Asset/Font.hpp>
+
+#include <Torchlight/Util/FontProcessor.hpp>
 
 void UICreationToolApplication::Start() {
 	m_wWindow = Window(m_strApplicationName, WINDOWED, Vec2F(800, 600), Vec2F(0, 0));
@@ -46,8 +46,8 @@ void UICreationToolApplication::Run() {
 
 	//FontProcessor::ProcessFont("./Assets/Fonts/JetBrainsMono-Bold.ttf");
 	//FontProcessor::ProcessFont("./Assets/Fonts/Sniglet.ttf");
-	//FontProcessor::ProcessFont("./Assets/Fonts/RobotoSlab-Bold.ttf");
-	FontProcessor::ProcessFont("./Assets/Fonts/calibri.ttf");
+	FontProcessor::ProcessFont("./Assets/Fonts/RobotoSlab-Bold.ttf");
+	//FontProcessor::ProcessFont("./Assets/Fonts/calibri.ttf");
 	//FontProcessor::ProcessFont("./Assets/Fonts/Envy Code R.ttf");
 
 	while (!m_wWindow.CloseRequested()) {
